Update README.md
OpenWeatherMap it not very privacy friendly and since it requiring a creditcard, it should (in my humble opinion) not be recommended.

Open Meteo is the leading data source for many weather services like OpenWeatherMap. It requires no API key or user information. 
It's the default weather integration for good reasons.